Added new files for updated project
The new file SNN_Navigation.ino is a simple test program using a 4x4 spiking neural network (SNN) for a simple navigation task. The algorithm used is the one described in Path Planning using a Spiking Neuron Algorithm with Axonal Delays by Jeffrey L. Krichmar. SNN_0_0_0_1.h is a simple library that implements this algorithm. Timer.h is the implementation of a non-blocking delay() (function in the Arduino library).